+++ dp/wifi3.0/be/dp_be_tx.h
@@ -142,7 +142,7 @@ void dp_tx_update_bank_profile(struct dp_soc_be *be_soc,
  * Return: QDF_STATUS_SUCCESS - success, others - failure
  */
 QDF_STATUS dp_tx_desc_pool_init_be(struct dp_soc *soc,
-				   uint16_t num_elem,
+				   uint32_t num_elem,
 				   uint8_t pool_id);
 /**
  * dp_tx_desc_pool_deinit_be() - De-initialize Tx Descriptor pool(s)
